requirements.txt
aiohttp==3.9.5
aiolimiter==1.1.0
aiosignal==1.3.1
annotated-types==0.7.0
async-timeout==4.0.3
attrs==23.2.0
contourpy==1.2.1
cycler==0.12.1
factor_analyzer==0.5.1
fonttools==4.53.0
frozenlist==1.4.1
idna==3.7
importlib_resources==6.4.0
joblib==1.4.2
json5==0.9.25
jsonlines==4.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
matplotlib==3.9.0
multidict==6.0.5
numpy==2.0.0
packaging==24.1
pandas==2.2.2
patsy==0.5.6
pillow==10.4.0
pydantic==2.7.4
pydantic_core==2.18.4
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
regex==2024.5.15
rpds-py==0.18.1
scikit-learn==1.5.0
scipy==1.13.1
seaborn==0.13.2
six==1.16.0
statsmodels==0.14.2
tenacity==8.4.2
threadpoolctl==3.5.0
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
yarl==1.9.4
zipp==3.19.2
